Remove unused code: support for storing part of index column in the value part of RocksDB key-value pair.
Initially, this space was used to store lengths of varchar columns
(mem-comparable images of varchars were endspace-padded). Then, we
switched to using mem-comparable encoding of varchars that stored
length in the key part of the key-value pair, so support for index
"tails" became dead code.

I'm removing it now as I'm going to make changes in data storage
(Issue #14), so I will likely break the dead code anyway.